Refactor syllable code
